Added selective destroyal of snapshots by user properties.
Individual datasets can now be marked with user properties that controls which types of snapshots that should always be destroyed. There are five different znap user properties.
  script.znap:nosnapshots
  script.znap:nomonthly
  script.znap.noweekly
  script.znap:nodaily
  script.znap:nohourly

If a dataset is marked with script.znap:noweekly=1, then weekly snapshots for that dataset will always be destroyed. This only affects snapshots taken by znap.